TRO-762 Adding classes for source and target
I want to make a common function for sync folders. If I do it with the current status of the class I need to pass a lot of arguments.

I want to have a source and target class so I can store some things in that classes so the functions calls are easier.

This commit is adding this classes. But not modifiying anything else neither using them

TRO-762 Adding classes for source and target

I want to make a common function for sync folders. If I do it with the current status of the class I need to pass a lot of arguments.

I want to have a source and target class so I can store some things in that classes so the functions calls are easier.

This commit is adding this classes. But not modifiying anything else neither using them