TRO-762 move the copy alert logic inside a class and solve existing folder
When we are copying the rules we have one problem. If the folder with the same name exists then we are failing.

With this PR I'm doing 2 things at once. Move everything inside the class. And also solving the problem. What I do instead of having a map with the id of the folder and the alert i just have a list of folders and a list of alerts. When the folder exists then I look if is the scenario fo the problem and then update the folder id in the alert